# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "Dokumentasi API e-commerce MK3",
        "contact": {
            "email": "kamu@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/detail-transactions": {
            "get": {
                "tags": [
                    "Detail Transactions"
                ],
                "summary": "Get all detail transactions (Admin only)",
                "operationId": "9c1687f675014b8aacae056ff861a53a",
                "responses": {
                    "200": {
                        "description": "List of detail transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Detail Transactions"
                ],
                "summary": "Create a new detail transaction (Admin only)",
                "operationId": "1922ba19ae1189f7702a9763e104cf27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_id",
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detail transaction created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/detail-transactions/{transaction_id}": {
            "get": {
                "tags": [
                    "Detail Transactions"
                ],
                "summary": "Get details of a specific transaction (Admin only)",
                "operationId": "747c18a8f5e92964cd5c8d04b0bfa188",
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "description": "Transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail transaction found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/detail-transactions/{id}": {
            "put": {
                "tags": [
                    "Detail Transactions"
                ],
                "summary": "Update a detail transaction quantity (Admin only)",
                "operationId": "2aef912f0bb47e8d5de0eba43b56e573",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Detail transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detail transaction updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Detail Transactions"
                ],
                "summary": "Delete a detail transaction (Admin only)",
                "operationId": "e6651629ea53c263fc342e3e631c9308",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Detail transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail transaction deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Detail transaction deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer Token via Sanctum",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}